Kd refactor
Reafactor usecase layer
