+++ app/usecase/interactor/tasklist_usecase.go
@@ -1,19 +1,20 @@
-package tasklistservice
+package interactor
 
 import (
 	entity "github.com/thvinhtruong/legoha/app/domain/entities"
+	repository "github.com/thvinhtruong/legoha/app/usecase/repository"
 )
 
 type TaskService struct {
-	taskRepo TaskListRepository
+	taskRepo repository.TaskListRepository
 }
 
-func NewTaskService(taskRepo TaskListRepository) *TaskService {
+func NewTaskService(taskRepo repository.TaskListRepository) *TaskService {
 	return &TaskService{taskRepo: taskRepo}
 }
 
 func (s *TaskService) AssignTask(user_id, todo_id int) error {
-	t := entity.TaskList{UserID: user_id, TodoID: todo_id, Finished: false}
+	t := entity.TodoUser{UserID: user_id, TodoID: todo_id, Finished: false}
 	err := s.taskRepo.Assign(&t)
 	if err != nil {
 		return err
@@ -22,7 +23,7 @@ func (s *TaskService) AssignTask(user_id, todo_id int) error {
 }
 
 func (s *TaskService) ListUsersTasks(user_id int) ([]*entity.User, error) {
-	t := entity.TaskList{UserID: user_id, Finished: false}
+	t := entity.TodoUser{UserID: user_id, Finished: false}
 	users, err := s.taskRepo.ListUsersForOneTodo(&t)
 	if err != nil {
 		return nil, err
@@ -31,7 +32,7 @@ func (s *TaskService) ListUsersTasks(user_id int) ([]*entity.User, error) {
 }
 
 func (s *TaskService) CompletedTask(user_id, todo_id int) error {
-	t := entity.TaskList{UserID: user_id, TodoID: todo_id, Finished: false}
+	t := entity.TodoUser{UserID: user_id, TodoID: todo_id, Finished: false}
 	err := s.taskRepo.Completed(&t)
 	if err != nil {
 		return err
@@ -40,7 +41,7 @@ func (s *TaskService) CompletedTask(user_id, todo_id int) error {
 }
 
 func (s *TaskService) RevokeTask(user_id, todo_id int) error {
-	t := entity.TaskList{UserID: user_id, TodoID: todo_id, Finished: false}
+	t := entity.TodoUser{UserID: user_id, TodoID: todo_id, Finished: false}
 	err := s.taskRepo.Revoke(&t)
 	if err != nil {
 		return err
@@ -49,7 +50,7 @@ func (s *TaskService) RevokeTask(user_id, todo_id int) error {
 }
 
 func (s *TaskService) UndoTask(user_id, todo_id int) error {
-	t := entity.TaskList{UserID: user_id, TodoID: todo_id, Finished: false}
+	t := entity.TodoUser{UserID: user_id, TodoID: todo_id, Finished: false}
 	err := s.taskRepo.Undo(&t)
 	if err != nil {
 		return err
