+++ app/usecase/repository/todo_reposity.go
@@ -1,4 +1,4 @@
-package todoservice
+package repository
 
 import (
 	entity "github.com/thvinhtruong/legoha/app/domain/entities"
@@ -10,12 +10,4 @@ type TodoRepository interface {
 	GetTodoByID(id int) (*entity.Todo, error)
 	PatchTodo(id int, t *entity.Todo) error
 	DeleteTodo(todo *entity.Todo) error
-}
-
-type TodoUseCase interface {
-	PostNewTodo(title, description string) error
-	ShowAllTodos() ([]*entity.Todo, error)
-	ShowTodoByID(id int) (*entity.Todo, error)
-	UpdateTodoInfor(id int, t *entity.Todo) error
-	DeleteTodo(id int) error
 }
