+++ app/usecase/repository/user_reposity.go
@@ -1,10 +1,10 @@
-package userusecase
+package repository
 
 import (
 	entity "github.com/thvinhtruong/legoha/app/domain/entities"
 )
 
-type Repository interface {
+type UserRepository interface {
 	CreateUser(entityUser *entity.User) error
 	ListUsers() ([]*entity.User, error)
 	GetUserByID(id int) (*entity.User, error)
@@ -12,13 +12,4 @@ type Repository interface {
 	PatchUser(id int, u *entity.User) error
 	DeleteUser(u *entity.User) error
 	LoginUser(user *entity.User) (*entity.User, error)
-}
-
-type UserUseCase interface {
-	RegisterUser(name, username, password string) error
-	FindAllUsers() ([]*entity.User, error)
-	FindUserById(id int) (*entity.User, error)
-	PatchUserInfor(id int, user *entity.User) error
-	DeleteUser(id int) error
-	LoginUser(username, password string) (*entity.User, error)
 }
